review issues fix
